We learn about arrow function, how arrow function is more readable,clean up function expression syntax and make code more structured. Arrow functions introduce concise body syntax, or implicit return. This allows the omission of the curly brackets and the return keyword. Implicit return is useful for creating succinct one-line operations in map , filter , and other common array methods.
Here () => means is that left part denotes the input of a function and the right part the output of that function.

Benefits of Arrow Functions
There are two major benefits of using Arrow functions. One is that it's a shorter syntax and thus requires less code. The main benefit is that it removes the several pain points associated with the this operator.

Arrow function run just fastere then normal funciton.
JS Rule — A JavaScript function will always return 'undefined' unless explicitly or implicitly told otherwise.

Note - Passing an arrow function in render will cause React to create a new function each time the component renders. But you don't need to worry because the impact on performance is very minuscule. It's very unlikely that your users will experience a noticeable lag just because of arrow functions in render.

(Ref - freecodecamp, stackoverflow,medium,sabhastian/react-arrow-function ...)